fix: display tag when version command is run
