+++ .dockerignore
@@ -1,6 +1,5 @@
 .github/
 .idea/
-.git/
 build/
 dist/
 node_modules/
